Added my current blueprints
Some Notes

====week1 collectables for exit====
CaveDrawing_Collectable_BP
-To change sprite, goto viewport then on right, look under "Sprite"

Exit_BP
-Same steps to change sprite.
-To set opening height, goto "Event Graph", double click "tlOpenExit"
then modify the value of the diamond on the top right

2DSideScrollerGameMode
-"Open the Exit"

=====week2 multiplayer====
2DSideScrollerCharacter
-"Event Graph" -> "Toggle Player 2"
-If problems with colliding with other character, ask me or mess with
2dsidscrollercharacter collision settings

2DSideScrollerGameMode
-"Toggle Player 2"

2DSideScrollerExampleMap
-"Set Up Camera for Multiplayer"
-"Update Multiplayer Camera Position"

====week3 damage====(such a mess)
2DSideScrollerGameMode
-"Check for Player Death"

You'll most likely notice when all players are dead, before they
respawn, the camera will angle to the side. We can probably fill this in
with a gameover or some other screen
Remember, to change the default sprite of a blueprint open it,
"ViewPort" tab, under "Sprite".
If you want multiple sprites for a blueprint, either duplicate it. Or
drag it in the scene as the "wrong" sprite, then in the Scene or "World
Outliner" select it then select either "PaperSprite" or
"RenderComponent" then under "Sprite"